[REMOVE]: ✂️ LocalDBUseCase 파일 분리 및 삭제
